# openapi/polaris-mail.yaml
openapi: 3.1.0
info:
  title: polaris-mail API
  version: 0.2.0
  description: |
    HMAC-authenticated REST surface for polaris-mail. All requests are signed with
    `X-Polaris-Sig: <hex(HMAC-SHA256(secret, canonical))>` where canonical is
    `polaris-api\n<METHOD>\n<path>\n<canonical-query>\n<X-Polaris-Ts>\n<X-Polaris-Nonce>\n<sha256-hex(body)>`.

    Webhooks delivered to subscribers (the WebhookEnvelope schema below) are signed
    `X-Polaris-Sig: <hex>` over the canonical string
    `polaris-webhook\nPOST\n<path>\n<canonical-query>\n<ts>\n<nonce>\n<sha256-hex(body)>`.
    Consumers should verify with the SDK (`@polaris-mail/sdk-node`'s `verifyWebhook`
    or `polariswebhook.VerifyWebhook` from `packages/sdk-go`).
servers:
  - url: https://polaris-mail-api.workers.dev
security:
  - HMACAuth: []
components:
  securitySchemes:
    HMACAuth:
      type: apiKey
      in: header
      name: X-Polaris-Sig
  parameters:
    KeyId:
      name: X-Polaris-Key-Id
      in: header
      required: true
      schema: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$' }
    Ts:
      name: X-Polaris-Ts
      in: header
      required: true
      schema: { type: string }
    Nonce:
      name: X-Polaris-Nonce
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Idempotency keys are scoped per principal: the same key string used by
        two different principals never collides. Retries from the same
        principal with the same key return the original `message_id`. Keys are
        cached for 24h.
      schema: { type: string, pattern: '^[A-Za-z0-9_-]{8,128}$' }
  schemas:
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
    KeyScope:
      type: string
      enum:
        - send
        - 'messages:read'
        - 'imap_bridge:read'
        - 'admin:rotate'
        - 'admin:read'
        - 'admin:impersonate'
...
    WebhookEventType:
      type: string
      enum:
        - message.received
        - message.sent
        - message.delivered
        - message.bounced
        - message.failed
...
    IssueApiKeyRequest:
      type: object
      required: [mailbox_id]
      properties:
        mailbox_id: { $ref: '#/components/schemas/Ulid' }
        display_name: { type: string, minLength: 1, maxLength: 120 }
        sender_ids:
          type: array
          items: { $ref: '#/components/schemas/Ulid' }
        scopes:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/KeyScope' }
        rate_limit_per_min: { type: integer, default: 1000 }
    IssueApiKeyResponse:
      type: object
      required: [key_id, key_secret, prefix, created_at]
      properties:
        key_id: { type: string }
        key_secret: { type: string }
        prefix: { type: string }
        created_at: { type: integer, format: int64 }
    RotateRequest:
      type: object
      required: [mode]
      properties:
        mode: { type: string, enum: [planned, emergency] }
        reason: { type: string, maxLength: 500 }
        new_label: { type: string }
    RotateResponse:
      type: object
      properties:
        new_key_id: { type: string }
        new_key_secret: { type: string }
        prev_status: { type: string }
    CreateWebhookSubRequest:
      type: object
      required: [mailbox_id, url, kind, events]
      properties:
        mailbox_id: { $ref: '#/components/schemas/Ulid' }
        url: { type: string, format: uri }
        kind: { type: string, enum: [external, tailnet] }
        events:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/WebhookEventType' }
    BridgeCredentialsResponse:
      type: object
      required: [updates, deletions, mirror_version]
      properties:
        updates:
          type: array
          items:
            type: object
            required: [id, username, bcrypt_hash, allowed_senders, mirror_version]
            properties:
              id: { type: string }
              username: { type: string }
              bcrypt_hash: { type: string }
              allowed_senders:
                type: array
                items: { type: string }
              mirror_version: { type: integer, format: int64 }
        deletions:
          type: array
          items: { type: string }
        mirror_version: { type: integer, format: int64 }
paths:
...
  /v1/admin/api-keys:
    post:
      summary: Issue an api key bound to a mailbox principal
      parameters:
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueApiKeyRequest' }
      responses:
        '201':
          description: issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssueApiKeyResponse' }
  /v1/admin/api-keys/{id}/rotate:
    post:
      summary: Rotate an api key
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - in: query
          name: dry_run
          schema: { type: string, enum: ['1'] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RotateRequest' }
      responses:
        '200':
          description: rotated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RotateResponse' }
  /v1/admin/api-keys/{id}/revoke:
    post:
      summary: Revoke an api key
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RotateRequest' }
      responses:
        '200':
          description: revoked
  /v1/admin/webhook-subs:
    post:
      summary: Create a webhook subscription
      parameters:
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWebhookSubRequest' }
      responses:
        '201':
          description: created
  /v1/bridge/credentials:
    get:
      summary: Mail-bridge credential mirror
      description: |
        Returns the current set of active SMTP submission credentials plus
        the set of disabled credentials to remove. Polled by the
        mail-bridge to keep its local SQLite mirror current. (Renamed from
        `/v1/daemon/credentials` — the legacy alias is gone.)
      parameters:
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
        - in: query
          name: since
          schema: { type: integer }
      responses:
        '200':
          description: credential delta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BridgeCredentialsResponse' }
  /v1/admin/bootstrap:
    post:
      summary: One-time seed of the first admin key. Signed with POLARIS_SECRET_A.
      security: []
      parameters:
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: seeded
        '409':
          description: already consumed
...